requirements.txt
# ── RAG Pipeline: PDF/DOCX → ChromaDB → Qwen/Mistral ──
pypdf                          # PDF parsing
python-docx                    # DOCX parsing
sentence-transformers          # local multilingual embeddings
chromadb                       # vector store
requests                       # Ollama HTTP API